refactor: lighten index.ts
