+++ frontend/weather-app/src/hooks/useSnackbar.ts
@@ -1,0 +1,26 @@
+// hooks/useSnackbar.ts
+import { useState } from "react";
+
+export const useSnackbar = () => {
+  const [snackbarOpen, setSnackbarOpen] = useState(false);
+  const [snackbarMessage, setSnackbarMessage] = useState("");
+  const [snackbarSeverity, setSnackbarSeverity] = useState<"success" | "error">(
+    "success",
+  );
+
+  const triggerSnackbar = (message: string, severity: "success" | "error") => {
+    setSnackbarMessage(message);
+    setSnackbarSeverity(severity);
+    setSnackbarOpen(true);
+  };
+
+  const handleSnackbarClose = () => setSnackbarOpen(false);
+
+  return {
+    snackbarOpen,
+    snackbarMessage,
+    snackbarSeverity,
+    triggerSnackbar,
+    handleSnackbarClose,
+  };
+};
