+++ frontend/weather-app/src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import CssBaseline from "@mui/material/CssBaseline";
 import { ThemeProvider } from "@mui/material/styles";
-import { AuthProvider } from "../providers/AuthContext";
-import { LocationProvider } from "../providers/LocationContext";
+import { AuthProvider } from "../providers/auth-context";
+import { LocationProvider } from "../providers/location-context";
 import theme from "../theme/theme";
 
 export default function MyApp({ Component, pageProps }: any) {
