+++ frontend/weather-app/src/pages/weather/forecast.tsx
@@ -9,8 +9,8 @@ import {
 import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
 import LoginModal from "../../components/login-modal";
-import { useAuth } from "../../providers/AuthContext";
-import { useLocation } from "../../providers/LocationContext";
+import { useAuth } from "../../providers/auth-context";
+import { useLocation } from "../../providers/location-context";
 import {
   CurrentTodayAndForecastsByDayDto,
   ForecastByDayDto,
